uart: rcar: Fix runtime configure
Calling configure at runtime was creating
unreadable data on the uart pins.

It appeared that disabling Receive and Transmit pins
at configuration as suggested by doc is not a thing
to do at runtime but only at initialization.

This commit is then moving this part to the init
function in order to able both a good initialization
phase and a well working "configure at runtime".

Signed-off-by: Aymeric Aillet <aymeric.aillet@iot.bzh>